Refactors application dashboard integration tests
These are now more in line with BDD. I have also removed the second test, which is now in the find by CRN test file.
